Update torrent model (#48)
* updated old entities and added category, subcategory and comment entity

* added like and updated comment entities

* updated repositories

* added UnitOfWork

* updated ViewModels

* added method PopularTorrents

* updated torrent repository

* added services

Category, subcategory, comment and like services

* added comment and like controllers

* updated comment service

* added Guard

* added search method for popular torrents

* added like service and updated view models

* updated like service

* added category and subcatetory controllers

+++ src/Rutracker.Client.Blazor/Services/AppState.cs
@@ -38,7 +38,7 @@ namespace Rutracker.Client.Blazor.Services
             return await IndexActionAsync(() => _torrentService.Torrents(page, pageSize, filter));
         }
 
-        public async Task<TorrentDetailsViewModel> Torrent(long id)
+        public async Task<TorrentDetailsViewModel> Torrent(int id)
         {
             return await IndexActionAsync(() => _torrentService.Torrent(id));
         }
